pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mboxer"
version = "0.1.0"
description = "Local-first MBOX archive processor for KM, RAG, and NotebookLM source packs."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "David G. Becker" }
]
dependencies = [
  "PyYAML>=6.0.1",
  "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5.0",
  "mypy>=1.10"
]
pdf = [
  "fpdf2>=2.7.9"
]

[project.urls]
Homepage = "https://github.com/xormania/mboxer"
Repository = "https://github.com/xormania/mboxer"
Issues = "https://github.com/xormania/mboxer/issues"

[project.scripts]
mboxer = "mboxer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mboxer = ["db/schema.sql", "db/migrations/*.sql"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
